+++ packages/solv/src/cli/update/jitoUpdate.ts
@@ -1,15 +1,11 @@
 import { spawnSync } from 'child_process'
 
-export const jitoUpdate = (tag: string, mod = false, isMajorThree = false, xdp = false) => {
-  // Update DZ
-  spawnSync(
-    `sudo apt install --only-upgrade doublezero doublezero-solana -y`,
-    { shell: true, stdio: 'inherit' },
-  )
-  spawnSync(
-    `sudo systemctl restart doublezerod`,
-    { shell: true, stdio: 'inherit' },
-  )
+export const jitoUpdate = (
+  tag: string,
+  mod = false,
+  isMajorThree = false,
+  xdp = false,
+) => {
   if (isMajorThree) {
     if (mod) {
       spawnSync(
@@ -64,7 +60,7 @@ export const jitoUpdate = (tag: string, mod = false, isMajorThree = false, xdp =
           stdio: 'inherit',
         },
       )
-      if(xdp) {
+      if (xdp) {
         spawnSync(
           `sudo setcap cap_net_raw,cap_net_admin,cap_bpf,cap_perfmon=p /home/solv/.local/share/solana/install/active_release/bin/agave-validator`,
           {
@@ -79,17 +75,17 @@ export const jitoUpdate = (tag: string, mod = false, isMajorThree = false, xdp =
       })
     }
   } else {
-    if(mod) {
+    if (mod) {
       spawnSync(
-          `sh -c "$(curl --netrc-optional -sSfL https://raw.githubusercontent.com/gabrielhicks/jito-solana/${tag}/installer)"`,
-          {
-            shell: true,
-            stdio: 'inherit',
-          },
-        )
+        `sh -c "$(curl --netrc-optional -sSfL https://raw.githubusercontent.com/gabrielhicks/jito-solana/${tag}/installer)"`,
+        {
+          shell: true,
+          stdio: 'inherit',
+        },
+      )
     } else {
       spawnSync(
-         `sh -c "$(curl --netrc-optional -sSfL https://release.jito.wtf/${tag}/install)"`,
+        `sh -c "$(curl --netrc-optional -sSfL https://release.jito.wtf/${tag}/install)"`,
         {
           shell: true,
           stdio: 'inherit',