+++ packages/solv/src/config/config.ts
@@ -56,11 +56,11 @@ export enum MAINNET_TYPES {
 
 // ⚠️ Please DO NOT forget to turn this to false if restart is not needed
 export const NODE_RESTART_REQUIRED_MAINNET = false
-export const NODE_RESTART_REQUIRED_TESTNET = true
+export const NODE_RESTART_REQUIRED_TESTNET = false
 export const FD_RESTART_REQUIRED_MAINNET = false
 export const FD_RESTART_REQUIRED_TESTNET = false
 export const BAM_RESTART_REQUIRED_MAINNET = false
-export const BAM_RESTART_REQUIRED_TESTNET = true
+export const BAM_RESTART_REQUIRED_TESTNET = false
 
 export type CONFIG_TYPE = {
   ID: string
@@ -97,6 +97,7 @@ export type CONFIG_TYPE = {
   ZERO_COPY: boolean
   JAG_SNAPSHOTS: boolean
   JAG_REGION: string
+  MEV_COMMISSION: number
 }
 
 export type PartialConfigType = Partial<CONFIG_TYPE>
@@ -137,6 +138,7 @@ export const CONFIG: CONFIG_TYPE = {
   ZERO_COPY: false,
   JAG_SNAPSHOTS: false,
   JAG_REGION: '',
+  MEV_COMMISSION: 1000,
 }
 
 export const HOME_PATHS = {